ClientePython-Plc/requirements.txt
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
cryptography==45.0.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.3.2
oauthlib==3.3.1
pycparser==2.22
pyqtgraph==0.13.7
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-snap7==2.0.2
repath==0.9.0
setuptools==80.9.0
shiboken6==6.9.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.1
